# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Huntly Admin Dashboard API",
    "description": "API for managing leads, clients, projects, finances, and internal projects. Use API Keys for automated integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "Huntly",
      "url": "https://huntlyai.xyz"
    }
  },
  "servers": [
    {
      "url": "https://admin.huntlyai.xyz/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development"
    }
  ],
  "security": [
    { "ApiKeyAuth": [] },
    { "CookieAuth": [] }
  ],
...
  "paths": {
...
    "/api-keys": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List all API keys",
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ApiKey" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Create a new API key",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateApiKeyRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyCreated" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for external integrations. Get one from the dashboard."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token",
        "description": "JWT token set after login"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Not authenticated",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "prefix": { "type": "string", "description": "Key prefix for identification" },
          "permissions": {
            "type": "array",
            "items": { "type": "string" }
          },
          "internalProjectId": { "type": "string", "nullable": true },
          "internalProject": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" }
            }
          },
          "lastUsedAt": { "type": "string", "format": "date-time", "nullable": true },
          "expiresAt": { "type": "string", "format": "date-time", "nullable": true },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateApiKeyRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "description": "Friendly name for the key" },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "transactions:read",
                "transactions:write",
                "transactions:delete",
                "internal-projects:read",
                "internal-projects:write",
                "tasks:read",
                "tasks:write",
                "tasks:delete",
                "full-access"
              ]
            },
            "default": ["transactions:read", "transactions:write"]
          },
          "internalProjectId": { "type": "string", "description": "Restrict key to specific project" },
          "expiresIn": {
            "type": "string",
            "enum": ["7d", "30d", "90d", "1y", "never"],
            "default": "never"
          }
        }
      },
      "ApiKeyCreated": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiKey" },
          {
            "type": "object",
            "properties": {
              "key": { "type": "string", "description": "The full API key - shown only once!" },
              "message": { "type": "string" }
            }
          }
        ]
      },
...
    }
  }
}